==> 4.gwas/02_gwas_continuous.Rmd <==
---
title: "GWAS - continuous trait"
author: "Filippo Biscarini"
date: "October 9, 2019"
output:
  html_document: default
  pdf_document: default
bibliography: gwas.bib
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library("knitr")
library("qqman")
library("dplyr")
library("ggplot2")
library("snpStats")
library("data.table")
library("doParallel")

source("../software/GWASfunction.R")

### own defined functions
qqPlot <- function(res) {
  
  q <- ggplot(res, aes(-log10( ppoints(length(P) )), -log10(sort(P,decreasing=F))))
  q <- q + geom_point() +  geom_abline(intercept=0,slope=1, col="red")
  q <- q + xlab( expression(Expected~~-log [10] (p)) ) + ylab( expression(Observed~~-log [10] (p)) ) 
  q <- q + ggtitle("")
  
  return(q)
}
```

## GWAS for continuous traits

We now perform a complete GWAS analysis for a continuous trait.
We use plant height (PH) from rice data published by @biscarini2016genome:

### Preparing the data

#### Genotypes

```{r data, echo=TRUE}
## Prepare data
pathD <- paste("../3.imputation/rice_imputed", c(".bed", ".bim", ".fam"), sep = "")
genotypes <- read.plink(pathD[1], pathD[2], pathD[3])

map <- genotypes$map
colnames(map) <- c("chr", "SNP", "gen.dist", "position", "A1", "A2")
```

N. **`r nrow(map)` SNPs** from **`r nrow(genotypes$genotypes)` rice samples**. SNPs are distributed over **`r length(unique(map$chr))` chromosomes** (chr: `r unique(map$chr)`); plant samples belong to **`r length(unique(genotypes$fam$pedigree))` populations** (populations: `r unique(genotypes$fam$pedigree)`)

#### Phenotypes

```{r, echo=TRUE}
## phenotypes
phenotypes = read.table("../data/rice_phenotypes.txt", header = TRUE)
names(phenotypes) <- c("id", "population", "PH")
```

Plant height (PH) was recorded on `r nrow(phenotypes)` rice samples. Genotype records and phenotype records need in the first place to be aligned:


```{r, echo=TRUE}
phenotypes <- phenotypes[phenotypes$id %in% row.names(genotypes$genotypes),]
row.names(phenotypes) <- phenotypes$id
phenotypes$id <- NULL
```

Data (both phenotypes and genotypes) are then saved to an object of class

```{r, echo=TRUE}
genData = list(SNP=genotypes$genotypes, MAP=map, PHENOTYPES=phenotypes)
save(genData, file = "genData_continuous.RData")

### analysis
load("genData_continuous.RData")
```

Below the distribution of phenotypes (PH) after alignment to genotypes:

```{r, echo=TRUE}
D <- genData$PHENOTYPES %>%
  group_by(population) %>%
  summarise(N=n(), avgPH=mean(PH), stdPH=sd(PH), minPH=min(PH),
            maxPH=max(PH))

kable(D)
```

```{r}
p <- ggplot(genData$PHENOTYPES, aes(x=PH, color=population, fill=population))
p <- p + geom_histogram(aes(y=..density..), alpha=0.5, position="identity")
p <- p + geom_density(alpha=.2) 
p
```

## Data filtering

```{r, echo=TRUE}
# Use SNP call rate of 100%, MAF of 0.1 (very stringent)
maf <- 0.05
callRate <- 1
SNPstats <- col.summary(genData$SNP)

maf_call <- with(SNPstats, MAF > maf & Call.rate == callRate)
genData$SNP <- genData$SNP[,maf_call]
genData$MAP <- genData$MAP[maf_call,]
SNPstats <- SNPstats[maf_call,]
```

## GWAS

### Set up

```{r, echo=TRUE}
# Choose trait for association analysis, use colnames(genData$LIP) for listing
# NOTE: Ignore the first column of genData$LIP (gender)
target <- "PH"
phenodata <- data.frame("id" = rownames(genData$PHENOTYPES),
                        "phenotype" = genData$PHENOTYPES[,target], stringsAsFactors = F)
trait_type = "gaussian"
```

```{r, echo=TRUE}
kable(head(phenodata))
```

### Execution

```{r, echo=TRUE, warning=FALSE, message=FALSE}
# Conduct GWAS (will take a while)
start <- Sys.time()
GWAA(genodata = genData$SNP, phenodata = phenodata, family = trait_type, filename = paste(target, ".txt", sep = ""))
Sys.time() - start # benchmark
```

### Results

```{r, echo=TRUE}
GWASout <- read.table(paste(target, ".txt", sep = ""), header = T, colClasses = c("character", rep("numeric",4)))
# GWASout$type <- rep("typed", nrow(GWASout))
GWASout$Neg_logP <- -log10(GWASout$p.value)
GWASout <- merge(GWASout, genData$MAP[,c("SNP", "chr", "position")])
GWASout <- GWASout[order(GWASout$Neg_logP, decreasing = T),]

gwasResults <- GWASout[,c("SNP","chr","position","p.value","t.value")]
names(gwasResults) <- c("SNP","CHR","BP","P","zscore")

kable(head(gwasResults))
```

#### Manhattan plot

```{r, echo=TRUE}
manhattan(gwasResults)
```

#### QQ plot

```{r, echo=TRUE}
qq(gwasResults$P)
```

Using an own-defined function:

```{r, echo=TRUE}
qqPlot(gwasResults)
```

#### Inflation factor $\lambda$

```{r, echo=TRUE}
## calculates lambda
z=qnorm(gwasResults$P/2)
lambda = round(median(z^2,na.rm=T)/qchisq(0.5,df=1),3)
print(lambda)
```

## References
